utils: move non-vscode-dependent utils to utils/
Refactor use of killProcessTree to a single function in utils/processUtils
