+++ src/goCover.ts
@@ -11,7 +11,7 @@ import path = require('path');
 import util = require('util');
 import vscode = require('vscode');
 import { isModSupported } from './goModules';
-import { envPath } from './goPath';
+import { envPath } from './utils/goPath';
 import { getTestFlags, goTest, showTestOutput, TestConfig } from './testUtils';
 import { getBinPath, getCurrentGoPath, getGoConfig, getWorkspaceFolderPath } from './util';
